Improve debug render config API and docs
