+++ src/plugins/debug/mod.rs
@@ -1,19 +1,29 @@
-//! Renders physics objects and properties for debugging purposes. This includes [AABBs](ColliderAabb),
-//! [colliders](Collider) and [contacts](Collision).
+//! Renders physics objects and properties for debugging purposes.
 //!
 //! See [`PhysicsDebugPlugin`].
 
+mod configuration;
 mod renderer;
 
+pub use configuration::*;
 pub use renderer::*;
 
 use crate::prelude::*;
 use bevy::prelude::*;
 
-/// Renders physics objects and properties for debugging purposes. This includes [AABBs](ColliderAabb),
-/// [colliders](Collider) and [contacts](Collision).
+/// Renders physics objects and properties for debugging purposes.
 ///
-/// You can use the [`PhysicsDebugConfig`] resource for the global configuration and the [`DebugRender`] component
+/// Currently, the following are supported for debug rendering:
+///
+/// - Entity axes
+/// - [AABBs](ColliderAabb)
+/// - [Collider] wireframes
+/// - [Contact] points
+/// - [Joints](joints)
+/// - Changing the visibility of entities to only show debug rendering
+///
+/// By default, only axes, colliders and joints are debug rendered. You can use the [`PhysicsDebugConfig`]
+/// resource for the global configuration and the [`DebugRender`] component
 /// for entity-level configuration.
 pub struct PhysicsDebugPlugin {
     schedule: Box<dyn ScheduleLabel>,
@@ -64,145 +74,6 @@ impl Plugin for PhysicsDebugPlugin {
                 )
                     .after(PhysicsSet::StepSimulation),
             );
-    }
-}
-
-/// Controls the global [`PhysicsDebugPlugin`] configuration.
-///
-/// To configure the debug rendering of specific entities, use the [`DebugRender`] component.
-#[derive(Reflect, Resource)]
-#[reflect(Resource)]
-pub struct PhysicsDebugConfig {
-    /// The lengths of the axes drawn for an entity.
-    pub axis_lengths: Option<Vector>,
-    /// The color of the [AABBs](ColliderAabb). If `None`, the AABBs will not be rendered.
-    pub aabb_color: Option<Color>,
-    /// The color of the [collider](Collider) wireframes. If `None`, the colliders will not be rendered.
-    pub collider_color: Option<Color>,
-    /// The color of the contact points. If `None`, the contact points will not be rendered.
-    pub contact_color: Option<Color>,
-    /// The color of the lines drawn from the centers of bodies to their joint anchors.
-    pub joint_anchor_color: Option<Color>,
-    /// The color of the lines drawn between joint anchors, indicating the separation.
-    pub joint_separation_color: Option<Color>,
-    /// Determines if the visibility of entities with [colliders](Collider) should be set to `Visibility::Hidden`,
-    /// which will only show the debug renders.
-    pub hide_meshes: bool,
-}
-
-impl Default for PhysicsDebugConfig {
-    fn default() -> Self {
-        Self {
-            #[cfg(feature = "2d")]
-            axis_lengths: Some(Vector::new(5.0, 5.0)),
-            #[cfg(feature = "3d")]
-            axis_lengths: Some(Vector::new(0.5, 0.5, 0.5)),
-            aabb_color: None,
-            collider_color: Some(Color::ORANGE),
-            contact_color: None,
-            joint_anchor_color: Some(Color::PINK),
-            joint_separation_color: Some(Color::RED),
-            hide_meshes: false,
-        }
-    }
-}
-
-/// A component for the debug render configuration of an entity.
-///
-/// This overwrites the global [`PhysicsDebugConfig`] for this specific entity.
-#[derive(Component, Reflect, Clone, Copy, PartialEq)]
-#[reflect(Component)]
-pub struct DebugRender {
-    /// The lengths of the axes drawn for the entity.
-    pub axis_lengths: Option<Vector>,
-    /// The color of the [AABB](ColliderAabb). If `None`, the AABB will not be rendered.
-    pub aabb_color: Option<Color>,
-    /// The color of the [collider](Collider) wireframe. If `None`, the collider will not be rendered.
-    pub collider_color: Option<Color>,
-    /// Determines if the entity's visibility should be set to `Visibility::Hidden`, which will only show the debug render.
-    pub hide_mesh: bool,
-}
-
-impl Default for DebugRender {
-    fn default() -> Self {
-        Self {
-            #[cfg(feature = "2d")]
-            axis_lengths: Some(Vector::new(5.0, 5.0)),
-            #[cfg(feature = "3d")]
-            axis_lengths: Some(Vector::new(0.5, 0.5, 0.5)),
-            aabb_color: None,
-            collider_color: Some(Color::LIME_GREEN),
-            hide_mesh: false,
-        }
-    }
-}
-
-impl DebugRender {
-    /// Creates a [`DebugRender`] configuration with the given lengths for the axes
-    /// that are drawn for the entity.
-    pub fn axes(axis_lengths: Vector) -> Self {
-        Self {
-            axis_lengths: Some(axis_lengths),
-            ..default()
-        }
-    }
-
-    /// Creates a [`DebugRender`] configuration with a given collider color.
-    pub fn collider(color: Color) -> Self {
-        Self {
-            collider_color: Some(color),
-            ..default()
-        }
-    }
-
-    /// Creates a [`DebugRender`] configuration with a given AABB color.
-    pub fn aabb(color: Color) -> Self {
-        Self {
-            aabb_color: Some(color),
-            ..default()
-        }
-    }
-
-    /// Sets the lengths of the axes drawn for the entity.
-    pub fn with_axes(mut self, axis_lengths: Vector) -> Self {
-        self.axis_lengths = Some(axis_lengths);
-        self
-    }
-
-    /// Sets the collider color.
-    pub fn with_collider_color(mut self, color: Color) -> Self {
-        self.collider_color = Some(color);
-        self
-    }
-
-    /// Sets the AABB color.
-    pub fn with_aabb_color(mut self, color: Color) -> Self {
-        self.aabb_color = Some(color);
-        self
-    }
-
-    /// Sets the visibility of the entity's visual mesh.
-    pub fn with_mesh_visibility(mut self, is_visible: bool) -> Self {
-        self.hide_mesh = !is_visible;
-        self
-    }
-
-    /// Disables axis debug rendering.
-    pub fn without_axes(mut self) -> Self {
-        self.axis_lengths = None;
-        self
-    }
-
-    /// Disables collider debug rendering.
-    pub fn without_collider(mut self) -> Self {
-        self.collider_color = None;
-        self
-    }
-
-    /// Disables AABB debug rendering.
-    pub fn without_aabb(mut self) -> Self {
-        self.aabb_color = None;
-        self
     }
 }
